oci 2019 exercises finished and bash file added

+++ Practice/oci/2019/start.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 g++ $1 -o program.out
-clear
-./program.out
-rm program.out
+echo "Continue? (y/n) > "
+read answer
+if [ "$answer" = "y" ]; then
+    clear
+    ./program.out
+    rm program.out
+else
+    echo "Exiting..."
+fi
